Tested install of the python virturalenv SRVr
Here is the script to install Python 3.6 with the virturalenv.  

Here is the what is included as well.  
1.  Google Cloud; use a Centos 7 OS made in the Compute Engine tab.
2.  Select cpu that is mini-cpu.  
3.  allow http and https
4.  Storage 10 Gb.  harddisk.
5. press create
6.  goto top left corner of gui console and useing the three horzontal bar graphic click on it to reveal the drop down.
7.  select VPC NETWORK AND THEN FIREWALL RULES.
8.  select from the center of the gui consoles header section in blue, the create firewall rule.
9.  The create firewall rule tab,  name your firewall same as your virtual OS instance that you are going to install python 3.6 on.
10.  Priority is 1000
11.  Network traffic select All instances on the Network.
12.   Address to use,  type in 0.0.0.0/0 (for all instances on the Network except port numbers and their intended use the protocol)
13.  Check TCP
14.  TCP port is 8000 make it so.
14.  Click create firewalll button (see the output for 0.0.0.0/0 and the protocol and port use)

Now you are ready after the created instance of your OS and the created firewall rules are complete.  
You may begin to install the script code of the python 3.6 to get a virtualenv service as the python server.  
When done your python 3.6 server will be running and ready to use.  use ctrl+z to shut it off.