I Followed the direction as said,
That there is no need to echo a message into a file.  Then add the file to the reposistory.  By use of a commit and a message about it.  
For this assignment.  

I used my own spun up Centos 7 machine from Google Cloud.  I added a repository to this assignment Called Personal-Repo-Site  and made directory for it in my home directory; /Personal-Repo-Site/.  I then echoed a README.md file to it.  About Terminator the 1984 movie scene.  Where Terminator telephone calls Sara Connors home.